Add support for 4-band Trx frontend
This also removes the code for I2C and serial Alex and Penelope/Pennylane
support on the C side, as it won't be needed with HAMLAB. GPIO code has been
left in, as it might be needed for key and microphone support on the HAMLAB box.

This has not yet been tested with the actual hardware!